Material Date Picker code restructure (#70708)
Date Picker restructuring. Moved files out of 'pickers' and merged several of the non-public implementation files into date_picker.dart.
