+++ packages/flutter/lib/material.dart
@@ -37,6 +37,7 @@ export 'src/material/button_bar_theme.dart';
 export 'src/material/button_style.dart';
 export 'src/material/button_style_button.dart';
 export 'src/material/button_theme.dart';
+export 'src/material/calendar_date_picker.dart';
 export 'src/material/card.dart';
 export 'src/material/card_theme.dart';
 export 'src/material/checkbox.dart';
@@ -51,6 +52,9 @@ export 'src/material/curves.dart';
 export 'src/material/data_table.dart';
 export 'src/material/data_table_source.dart';
 export 'src/material/data_table_theme.dart';
+export 'src/material/date.dart';
+export 'src/material/date_picker.dart';
+export 'src/material/date_picker_deprecated.dart';
 export 'src/material/debug.dart';
 export 'src/material/dialog.dart';
 export 'src/material/dialog_theme.dart';
@@ -82,6 +86,7 @@ export 'src/material/ink_ripple.dart';
 export 'src/material/ink_splash.dart';
 export 'src/material/ink_well.dart';
 export 'src/material/input_border.dart';
+export 'src/material/input_date_picker_form_field.dart';
 export 'src/material/input_decorator.dart';
 export 'src/material/list_tile.dart';
 export 'src/material/material.dart';
@@ -97,7 +102,6 @@ export 'src/material/outlined_button_theme.dart';
 export 'src/material/page.dart';
 export 'src/material/page_transitions_theme.dart';
 export 'src/material/paginated_data_table.dart';
-export 'src/material/pickers/pickers.dart';
 export 'src/material/popup_menu.dart';
 export 'src/material/popup_menu_theme.dart';
 export 'src/material/progress_indicator.dart';
