--- packages/flutter/lib/src/material/pickers/pickers.dart
@@ -1,21 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-// Date Picker public API
-export 'calendar_date_picker.dart' show CalendarDatePicker;
-export 'date_picker_common.dart' show
-  DatePickerEntryMode,
-  DatePickerMode,
-  DateTimeRange,
-  SelectableDayPredicate;
-export 'date_picker_deprecated.dart';
-export 'date_picker_dialog.dart' show showDatePicker;
-export 'date_range_picker_dialog.dart' show showDateRangePicker;
-export 'input_date_picker.dart' show InputDatePickerFormField;
-
-// TODO(ianh): Not exporting everything is unusual and we should
-// probably change to just exporting everything and making sure it's
-// acceptable as a public API, or, worst case, merging the parts
-// that really must be public into a single file and make them
-// actually private.
